practice/resume_analyzer/requirements.txt
streamlit==1.49.1
pandas==2.3.2
python-dotenv==1.1.1
pypdf==6.0.0
docx2txt==0.9
requests==2.32.5